FlaskApp/requirements.txt
streamlit
tensorflow
numpy
opencv-python
pillow
matplotlib
flask